+++ tests/unit/Task/CodeceptionTest.php
@@ -44,14 +44,16 @@ class CodeceptionTest extends \Codeception\TestCase\Test
             ->getCommand()
         )->equals('codecept run unit Codeception/Command --group core --env process1 --coverage');
 
+        $failGroupName = 'failed1';
         verify((new \Robo\Task\Testing\Codecept('codecept'))
             ->test('tests/unit/Codeception')
             ->configFile('~/Codeception')
             ->xml('result.xml')
             ->html()
             ->noRebuild()
+            ->failGroup($failGroupName)
             ->getCommand()
-        )->equals('codecept run tests/unit/Codeception -c ~/Codeception --xml result.xml --html --no-rebuild');
+        )->regExp("|^codecept run tests/unit/Codeception -c ~/Codeception --xml result\\.xml --html --no-rebuild --override ['\"]extensions: config: Codeception\\\\Extension\\\\RunFailed: fail-group: {$failGroupName}['\"]$|");
 
         verify((new \Robo\Task\Testing\Codecept('codecept.phar'))->debug()->getCommand())->contains(' --debug');
         verify((new \Robo\Task\Testing\Codecept('codecept.phar'))->silent()->getCommand())->contains(' --silent');
@@ -59,6 +61,9 @@ class CodeceptionTest extends \Codeception\TestCase\Test
         verify((new \Robo\Task\Testing\Codecept('codecept.phar'))->tap()->getCommand())->contains('--tap');
         verify((new \Robo\Task\Testing\Codecept('codecept.phar'))->json()->getCommand())->contains('--json');
         verify((new \Robo\Task\Testing\Codecept('codecept.phar'))->noRebuild()->getCommand())->contains('--no-rebuild');
+        $failGroupName = 'failed2';
+        verify((new \Robo\Task\Testing\Codecept('codecept.phar'))->failGroup($failGroupName)->getCommand())
+            ->regExp("|--override ['\"]extensions: config: Codeception\\\\Extension\\\\RunFailed: fail-group: {$failGroupName}['\"]|");
     }
 
 }